refactor: extract clientprofile and requestinterceptor packages
Move campaign-matching types (Rule, RuleSet, Headers, GlobMap) from
pkg/telemetry/phonehome to pkg/clientprofile, breaking the coupling
between request parameter handling and client profile matching.

Move RequestParams, request detail extraction, and helper functions
to pkg/grpc/common/requestinterceptor, with RequestParams.Headers
now using http.Header directly.

Type aliases in phonehome preserve backward compatibility.

Partially AI-generated.
