+++ roxctl/common/client_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/stackrox/rox/pkg/clientconn"
+	"github.com/stackrox/rox/pkg/clientprofile"
 	"github.com/stackrox/rox/pkg/env"
-	"github.com/stackrox/rox/pkg/telemetry/phonehome"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -24,14 +24,14 @@ func TestSanitizeHeaderValue(t *testing.T) {
 
 func Test_setCustomHeaders(t *testing.T) {
 	headers := http.Header{}
-	setCustomHeaders(phonehome.Headers(headers).Set)
+	setCustomHeaders(clientprofile.Headers(headers).Set)
 	assert.Len(t, headers, 2)
 	assert.Equal(t, "", headers.Get(clientconn.RoxctlCommandHeader))
 	assert.Equal(t, "1", headers.Get(clientconn.RoxctlCommandIndexHeader))
 
 	t.Setenv(env.ExecutionEnvironment.EnvVar(), "test")
 	RoxctlCommand = "custom command"
-	setCustomHeaders(phonehome.Headers(headers).Set)
+	setCustomHeaders(clientprofile.Headers(headers).Set)
 	assert.Len(t, headers, 3)
 	assert.Equal(t, RoxctlCommand, headers.Get(clientconn.RoxctlCommandHeader))
 	assert.Equal(t, "2", headers.Get(clientconn.RoxctlCommandIndexHeader))
